Rename locale files for consistency
